Use the certificates directly from flash to reduce RAM usage
The original solution of loading the certificates from SPIFFS is very wasteful of RAM, 64k is a large chunk of the usable heap. As far as I can tell, if the the data is embedded with the embed_txtfiles approach (https://docs.platformio.org/en/latest/platforms/espressif32.html#embedding-binary-data), then it is read directly from flash through the cache mechanism (https://www.exploreembedded.com/wiki/File:Address_map.JPG) and the only RAM needed is the index that esp_crt_bundle_init builds.
It would be even better to just use the approach the original Espressif authors intended (see the EMBEDDED_BUNDLE == 1 section of esp_crt_bundle_attach), but that would be messier in the configuration due to the file path to variable name mapping.